+++ internal/models/menu.go
@@ -1,9 +1,9 @@
 package models
 
 type Menu struct {
-	ID          int64  `json:"id"`
-	SupplierID  int64  `json:"supplier"`
-	CategorieID int64  `json:"categorie"`
+	ID          int    `json:"id"`
+	SupplierID  int    `json:"supplier"`
+	CategorieID int    `json:"categorie"`
 	Image       string `json:"image"`
 	Price       int    `json:"price"`
 }
